+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Colombia region codes
+- Peru region codes
 
 ## [1.4.0] - 2022-08-08